+++ app/src/main/AndroidManifest.xml
@@ -10,9 +10,6 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".Home"
-            android:exported="true" />
-        <activity
             android:name=".UpdatePhone"
             android:exported="true" />
         <activity
