+++ app/src/main/java/com/example/mad/UpdatePhone.java
@@ -21,7 +21,7 @@ public class UpdatePhone extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_update_customer);
+        setContentView(R.layout.activity_update_phone);
 
         customerIDd = findViewById(R.id.customerIDdup);
         customerNamed = findViewById(R.id.customerNamedup);
