+++ app/src/main/res/layout/row.xml
@@ -45,7 +45,7 @@
                 app:layout_constraintTop_toTopOf="@+id/customer_id_txt" />
 
             <TextView
-                android:id="@+id/customer_category_txt"
+                android:id="@+id/customer_email_txt"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:text="Customer Name"
